+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #77 Use percentage metric for thread pool utilization check
 * #78 Metrics renaming to follow naming convention
 * #79 Health check renaming to follow naming convention
+* #80 JDK9 Compatibility
 
 ## 0.25