+++ RELEASE-NOTES.md
@@ -1,4 +1,4 @@
-## 0.26
+## 1.0.RC1
 
 * #74 Add a @BQConfigProperty for 'compactPath' in ContextHandler
 * #75 Cleaning up APIs deprecated since <= 0.25